♻️ Refactor: 페이지통신요청 -> 컴포넌트에서 통신요청
