+++ components/mypage/MyMeetings_mypage.tsx
@@ -1,14 +1,36 @@
 import { MeetingWrapper } from './mypage.style';
 import MyMeetingCard from './MyMeetingCard';
-import { useRecoilValue } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 import Link from 'next/link';
 import { attendingMeeting_opened, hostMeeting_opened } from '../../atoms/mypage/selector';
+import { useEffect } from 'react';
+import axios from 'axios';
+import { mypageAttendingMeetings, mypageHostMeetings } from '../../atoms/mypage/atoms';
 
 const MyMeetings_mypage = () => {
+   const API_URI = process.env.NEXT_PUBLIC_API_URI;
+   const setHostMeetings = useSetRecoilState(mypageHostMeetings);
+   const setAttendingMeetings = useSetRecoilState(mypageAttendingMeetings);
    const host_openedMeetings = useRecoilValue(hostMeeting_opened);
    const attending_openedMeetings = useRecoilValue(attendingMeeting_opened);
-   // const host_openedMeetings: any = hostMeetings.filter((meeting:any) => meeting.meetingStatus === true);
-   // const attending_openedMeetings: any = hostMeetings.filter((meeting:any) => meeting.meetingStatus === false);
+
+   useEffect(() => {
+      axios.defaults.headers.common['Authorization'] = localStorage.getItem('AccessToken');
+      // eslint-disable-next-line import/no-named-as-default-member
+      axios
+         .all([
+            axios.get(API_URI + '/mypage/meetings/hosts?page=1&size=20'),
+            axios.get(API_URI + '/mypage/meetings/participants?page=1&size=20')
+         ])
+         .then(
+            // eslint-disable-next-line import/no-named-as-default-member
+            axios.spread((res1, res2) => {
+               setHostMeetings(res1.data);
+               setAttendingMeetings(res2.data);
+            })
+         )
+         .catch(err => console.log(err));
+   }, []);
 
    return (
       <>
